requirements.txt
certifi==2019.3.9
easygui==0.98.1
mysql-connector-python==8.0.15
numpy==1.16.2
pandas==0.24.1
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
sockets==1.0.0
